% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wikit}
\alias{wikit}
\title{Search for Wikipedia summaries}
\usage{
wikit(topic, lang = "en", ncols = 140)
}
\arguments{
\item{topic}{Topic to search for}

\item{lang}{HTML ISO language code, e.g. "en"}

\item{ncols}{Number of columns}
}
\description{
Return character string of Wikipedia summary.
}
\examples{
library(outsider)
wikit <- module_import('wikit', repo = 'dombennett/om..wikit')
penguin_summary <- wikit('penguin')
cat(penguin_summary)
}
